requirements.txt
python==3.6.12
pytorch==1.3.1
